refactor(go): Extract digest functionality to separate file
